// frontend/src/api/generated/model/noteCreateRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * crossbill API
 * OpenAPI spec version: 0.0.0
 */
import type { NoteCreateRequestKind } from './noteCreateRequestKind.ts';

/**
 * Schema for creating a note.
 */
export interface NoteCreateRequest {
  /**
   * Note title
   * @minLength 1
   */
  title: string;
  /** Markdown body */
  body?: string;
  /** Optional note kind */
  kind?: NoteCreateRequestKind;
  /** Book this note is created in */
  book_id: number;
  chapter_ids?: number[];
  highlight_ids?: number[];
  tag_ids?: number[];
}